# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Vraja Content Hub public contracts and critical editorial mutations",
    "version": "1.1.0",
    "description": "Local implementation. Additive v1 fields. No deployment or Android integration claim. Complete admin workflow in API-CONTRACT.md."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local. Verify the approved staging/production hostname before use."
    }
  ],
  "paths": {
    "/api/v1/wallpaper-categories": {
      "get": {
        "summary": "Complete ancestor tree, including inactive and empty nodes. Pagination rejected.",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "description": "India day. Default today in Asia/Kolkata.",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful validated response. Only complete scope reconciliation can withdraw cached server-owned content.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryList"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request. Retain cached content and personal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "No eligible item or detail unavailable. Retain cached content and personal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Version or snapshot changed. Refresh and retry. Retain cached content and personal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Request failed. Retain cached content and personal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Dependency or schema unavailable. Retain cached content and personal state.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "details": {},
          "retryable": {
            "type": "boolean"
          }
        },
        "required": [
          "message"
        ],
        "additionalProperties": true
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "parent_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "active_wallpaper_count": {
            "type": "integer",
            "minimum": 0
          },
          "is_active": {
            "type": "boolean"
          },
          "is_selectable": {
            "type": "boolean"
          },
          "sort_order": {
            "type": "integer"
          },
          "editorial_is_active": {
            "type": "boolean",
            "description": "Own setting; is_active also includes all ancestors."
          }
        },
        "required": [
          "id",
          "name",
          "parent_id",
          "active_wallpaper_count",
          "is_active",
          "is_selectable",
          "sort_order"
        ],
        "additionalProperties": true
      },
...
      "CategoryList": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Category"
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}